+++ browser_plugins/customer/src/types/index.ts
@@ -1,0 +1,9 @@
+import { CustomerOptionType } from "@heimdallr-sdk/types";
+
+export enum CustomerTypes {
+  CUSTOMER = 111
+}
+
+export interface CustomerOptions {
+  customers?: CustomerOptionType[];
+}