+++ browser_plugins/console/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @heimdallr-sdk/console
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @heimdallr-sdk/types@0.0.18
+  - @heimdallr-sdk/utils@0.0.23
+
 ## 0.0.19
 
 ### Patch Changes