+++ browser_plugins/performance/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @heimdallr-sdk/performance
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @heimdallr-sdk/types@0.0.18
+  - @heimdallr-sdk/utils@0.0.23
+
 ## 1.0.0
 
 ### Major Changes