+++ browser_plugins/record/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @heimdallr-sdk/record
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @heimdallr-sdk/types@0.0.18
+  - @heimdallr-sdk/utils@0.0.23
+
 ## 0.0.7
 
 ### Patch Changes