+++ browser_plugins/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @heimdallr-sdk/vue
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @heimdallr-sdk/types@0.0.18
+  - @heimdallr-sdk/utils@0.0.23
+
 ## 0.0.23
 
 ### Patch Changes